# src/main/resources/v3/eric-oss-app-lcm-sdk-openapi.yaml
...
openapi: 3.0.2
info:
  x-api-id: 12f6a604-1712-43a6-a175-467bbcc0327e
  x-audience: external-partner
  description: |
    The App Life Cycle Management (LCM) API controls and monitors the life cycle of the Apps and their instances.
  version: 3.0.0
  title: App Lifecycle Management API
  contact:
    url: https://www.ericsson.com
...
servers:
  - url: https://{eic-host}/app-lifecycle-management/v3
    variables:
      host:
        default: localhost
        description: Change this value to point to your custom host.
paths:
...
  /app-instances:
    get:
      tags:
        - App Instances
      summary: Get App instances
      description: Get a list of all available App instances. This endpoint allows users to fetch information about multiple App instances present in the system, providing an overview of the available instances and their details.
      operationId: getAppInstances
      parameters:
        - name: appId
          in: query
          description: "App Id to get instance information"
          required: false
          schema:
            type: string
          examples:
            appId:
              summary: Example of App resource identifier
              value: 26471a81-1de4-4ad9-9724-326eefd22230
        - name: offset
          in: query
          description: |
            To get a paginated list of App Instances, use the offset and limit query params to manage the number of App Instances to returned.
            The offset specifies the start element for the App Instances returned. To get the next page of entries, the API needs to be called with an offset
            parameter that equals the sum of the previous offset value and limit returned in the previous result.
          required: false
          explode: true
          schema:
            type: string
            minimum: 0
          examples:
            offsetExample:
              summary: Example of Offset
              value: /v3/app-instances?offset=11&limit=10
        - name: limit
          in: query
          description: |
            Limit is used with the offset query param to return a paginated list of App Instances. The limit specifies the max number of App Instances to return.
          required: false
          style: form
          explode: true
          schema:
            type: string
            minimum: 1
          examples:
            pageLimitExample:
              summary: Example of Limit
              value: /v3/app-instances?offset=11&limit=10
        - name: Accept
          description: Indicates the response media type accepted by the client.
          in: header
          required: false
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AppInstanceItems'
              examples:
                AppInstancesResponse:
                  $ref: '#/components/examples/AppInstancesResponse'
        "400":
          $ref: '#/components/responses/BadRequest'
        "404":
          $ref: '#/components/responses/NotFound'
        "500":
          $ref: '#/components/responses/InternalServerError'
    post:
      tags:
        - App Instances
      summary: Create an App Instance
      description: This endpoint initiates the instance creation of an App, coordinating a series of verification and setup steps. Finally, it generates instance property files and sends an instance creation request to the ACM.  Upon completion, it assigns a unique identifier to track the process and stores the instance details in the LCM database.
      operationId: "createAppInstance"
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateAppInstanceRequest'
            examples:
              CreateAppInstanceRequest:
                $ref: '#/components/examples/CreateAppInstanceRequest'
        required: true
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AppInstance'
              examples:
                CreateAppInstanceResponse:
                  $ref: '#/components/examples/CreateAppInstanceResponse'
        "400":
          $ref: '#/components/responses/BadRequest'
        "404":
          $ref: '#/components/responses/NotFound'
        "500":
          $ref: '#/components/responses/InternalServerError'
  /app-instances/{appInstanceId}:
    get:
      tags:
        - App Instances
      summary: Get an App instance based on instance Id
      description: Get detailed information about a specific App instance identified by its unique instance Id. This endpoint enables users to fetch specific details of an App instance using its unique identifier for detailed inspection and review.
      operationId: getAppInstanceById
      parameters:
        - in: path
          description: instanceId
          required: true
          name: appInstanceId
          schema:
            type: string
          examples:
            appInstanceId:
              summary: Example of App instance resource identifier
              value: 7e151de6-18a9-4770-be4f-354b620f0035
        - name: Accept
          description: Indicates the response media type accepted by the client.
          in: header
          required: false
          schema:
            type: string
      responses:
        "200":
          description: OK, App Instances response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AppInstance'
              examples:
                AppInstanceResponse:
                  $ref: '#/components/examples/AppInstanceResponse'
        "400":
          $ref: '#/components/responses/BadRequest'
        "404":
          $ref: '#/components/responses/NotFound'
        "500":
          $ref: '#/components/responses/InternalServerError'
    delete:
      tags:
        - App Instances
      summary: Delete a specific App instance
      description: Initiates the deletion of a specific instance of an application from the system. Upon initiation, the API responds with a status code 202 (Accepted), indicating that the request has been accepted for processing. Since this operation may take some time to complete, users can retrieve the app instance using the link in the response. They can then monitor the status until the app instance has been successfully deleted.
      operationId: deleteAppInstance
      parameters:
        - name: appInstanceId
          in: path
          description: Id of the App Instance to be deleted
          required: true
          style: simple
          explode: false
          schema:
            type: string
          examples:
            appInstanceId:
              summary: Example of App Instance resource identifier
              value: 7e151de6-18a9-4770-be4f-354b620f0035
        - name: Accept
          description: Indicates the response media type accepted by the client.
          in: header
          required: false
          schema:
            type: string
      responses:
        "202":
          description: Accepted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AppInstanceOperationResponse'
              examples:
                deleteAppInstanceResponse:
                  $ref: '#/components/examples/DeleteAppInstanceResponse'
        "400":
          $ref: '#/components/responses/BadRequest'
        "404":
          $ref: '#/components/responses/NotFound'
        "500":
          $ref: '#/components/responses/InternalServerError'
...
  /app-instances/{appInstanceId}/component-instances:
    put:
      tags:
        - App Instances
      summary: Update details of a specific App Instance
      description: Allows users to modify and update the details of a specific App instance identified by its unique instance Id.
      operationId: updateAppInstance
      parameters:
        - in: path
          description: instanceId
          required: true
          name: appInstanceId
          schema:
            type: string
          examples:
            appInstanceId:
              summary: Example of App instance resource identifier
              value: 7e151de6-18a9-4770-be4f-354b620f0035
        - name: Accept
          description: Indicates the response media type accepted by the client.
          in: header
          required: false
          schema:
            type: string
        - name: Content-Type
          description: This specifies the media type of the request sent by the client to the server.
          in: header
          required: false
          schema:
            type: string
          example: application/json
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateAppInstanceRequest'
            examples:
              UpdateAppInstanceRequest:
                $ref: '#/components/examples/UpdateAppInstanceRequest'
        required: true
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AppInstanceUpdateResponse'
              examples:
                UpdateAppInstanceResponseUndeployed:
                  $ref: '#/components/examples/UpdateAppInstanceResponseUndeployed'
        "202":
          description: Accepted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AppInstanceUpdateResponse'
              examples:
                UpdateAppInstanceResponse:
                  $ref: '#/components/examples/UpdateAppInstanceResponse'
        "400":
          $ref: '#/components/responses/BadRequest'
        "404":
          $ref: '#/components/responses/NotFound'
        "500":
          $ref: '#/components/responses/InternalServerError'
components:
  schemas:
...
    CreateAppInstanceRequest:
      type: object
      properties:
        appId:
          type: string
    UpdateAppInstanceRequest:
      type: object
      properties:
        componentInstances:
          type: array
          items:
            $ref: '#/components/schemas/ComponentInstances'
...
    Event:
      type: object
      properties:
        type:
          type: string
        title:
          type: string
        detail:
          type: string
        createdAt:
          type: string
    Href:
      type: object
      properties:
        href:
          type: string
...
    AppInstanceItems:
      type: object
      properties:
        items:
          description: The list of all App Instances
          type: array
          items:
            $ref: "#/components/schemas/AppInstance"
    AppInstance:
      type: object
      properties:
        id:
          type: string
        appId:
          type: string
        status:
          $ref: '#/components/schemas/AppInstanceStatus'
        createdAt:
          type: string
        updatedAt:
          type: string
        credentials:
          $ref: '#/components/schemas/Credentials'
        componentInstances:
          type: array
          items:
            $ref: '#/components/schemas/ComponentInstances'
        events:
          type: array
          items:
            $ref: '#/components/schemas/Event'
        self:
          $ref: '#/components/schemas/Href'
        app:
          $ref: '#/components/schemas/Href'
    Credentials:
      type: object
      properties:
        clientId:
          type: string
    ComponentInstances:
      type: object
      properties:
        name:
          type: string
        version:
          type: string
        type:
          type: string
        description:
          type: string
        deployState:
          type: string
        message:
          type: string
        properties:
          type: object
          description: "Property values associated with the Component Instance. Can be user defined or default properties."
    AppInstanceStatus:
      type: string
      enum:
        - UNDEPLOYED
        - UNDEPLOYING
        - UNDEPLOY_ERROR
        - DEPLOYED
        - DEPLOYING
        - DEPLOY_ERROR
        - DELETING
        - DELETE_ERROR
        - UPDATING
        - UPDATE_ERROR
        - UPGRADING
    AppInstanceOperationResponse:
      type: object
      properties:
        appInstance:
          type: object
          properties:
            status:
              $ref: '#/components/schemas/AppInstanceStatus'
            href:
              type: string
...
    AppInstanceUpdateResponse:
      type: object
      properties:
        componentInstances:
          type: array
          items:
            $ref: '#/components/schemas/ComponentInstances'
        appInstance:
          type: object
          properties:
            id:
              type: string
            href:
              type: string
    ProblemDetails:
      required:
        - title
        - status
        - detail
      type: object
      properties:
        title:
          type: string
          description: A short, human-readable summary of the problem type.
        status:
          type: integer
          description: The HTTP status code generated by the server
        detail:
          type: string
          description: A human-readable explanation specific to this occurrence of the problem.
  responses:
    BadRequest:
      description: Bad Request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ProblemDetails'
          examples:
            BadRequest:
              $ref: '#/components/examples/BadRequest'
    NotFound:
      description: Not Found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ProblemDetails'
          examples:
            NotFound:
              $ref: '#/components/examples/NotFound'
    InternalServerError:
      description: Internal Server Error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ProblemDetails'
          examples:
            InternalServerError:
              $ref: '#/components/examples/InternalServerError'
  examples:
...
    CreateAppInstanceRequest:
      summary: Create App Instance Request Body
      value:
        appId: "26471a81-1de4-4ad9-9724-326eefd22230"
...
    CreateAppInstanceResponse:
      summary: Create App Instance Response
      value:
        id: 7e151de6-18a9-4770-be4f-354b620f0035
        appId: 26471a81-1de4-4ad9-9724-326eefd22230
        status: 'UNDEPLOYED'
        createdAt: "2023-04-06T00:04:16.711+00:00"
        updatedAt: "2023-04-06T00:05:16.711+00:00"
        credentials:
          clientId: rappid-3146ccdc-0323-4f34-8f3e-13b858c1c582-1708549936654-8c12ffcc-64c3-4070-b9b5-a115ded1f825
        componentInstances:
          - name: "eric-oss-hello-metrics-go-app"
            version: "1.1.1"
            type: Microservice
            deployState: 'UNDEPLOYED'
            properties:
              timeout: 5
              userDefinedHelmParameters:
                replicaCount: 2
        events: []
        self:
          href: "app-lifecycle-management/v3/app-instances/7e151de6-18a9-4770-be4f-354b620f0035"
        app:
          href: "app-lifecycle-management/v3/apps/26471a81-1de4-4ad9-9724-326eefd22230"
...
    BadRequest:
      value:
        title: Bad Request
        status: 400
        detail: The message body is missing a required field.
    NotFound:
      value:
        title: Not Found
        status: 404
        detail: The requested URL is not found.
    InternalServerError:
      value:
        title: Internal Server Error
        status: 500
        detail: Internal Server Error Occurs.
    AppInstanceResponse:
      summary: App Instance Response
      value:
        id: 7e151de6-18a9-4770-be4f-354b620f0035
        appId: 26471a81-1de4-4ad9-9724-326eefd22230
        status: DEPLOYED
        createdAt: "2023-04-06T00:04:16.711+00:00"
        updatedAt: "2023-04-06T00:05:16.711+00:00"
        credentials:
          clientId: rappid-3146ccdc-0323-4f34-8f3e-13b858c1c582-1708549936654-8c12ffcc-64c3-4070-b9b5-a115ded1f825
        componentInstances:
          - name: "eric-oss-hello-metrics-go-app"
            version: "1.1.1"
            type: Microservice
            deployState: DEPLOYED
            properties:
              timeout: 5
              namespace: "namespace-example"
              userDefinedHelmParameters:
                replicaCount: 2
        events: []
        self:
          href: "app-lifecycle-management/v3/app-instances/7e151de6-18a9-4770-be4f-354b620f0035"
        app:
          href: "app-lifecycle-management/v3/apps/26471a81-1de4-4ad9-9724-326eefd22230"
    AppInstancesResponse:
      summary: App Instances Response
      value:
        items:
          - id: 7e151de6-18a9-4770-be4f-354b620f0035
            appId: 26471a81-1de4-4ad9-9724-326eefd22230
            status: DEPLOYED
            createdAt: "2023-04-06T00:04:16.711+00:00"
            updatedAt: "2023-04-06T00:05:16.711+00:00"
            credentials:
              clientId: rappid-3146ccdc-0323-4f34-8f3e-13b858c1c582-1708549936654-8c12ffcc-64c3-4070-b9b5-a115ded1f825
            componentInstances:
              - name: "eric-oss-hello-metrics-go-app"
                version: "1.1.1"
                type: Microservice
                deployState: DEPLOYED
                properties:
                  timeout: 5
                  namespace: "namespace-example"
                  userDefinedHelmParameters:
                    replicaCount: 2
            events: []
            self:
              href: "app-lifecycle-management/v3/app-instances/7e151de6-18a9-4770-be4f-354b620f0035"
            app:
              href: "app-lifecycle-management/v3/apps/26471a81-1de4-4ad9-9724-326eefd22230"
...
    DeleteAppInstanceResponse:
      summary: Delete an App Instance response
      value:
        appInstance:
          status: "DELETING"
          href: "app-lifecycle-management/v3/app-instances/7e151de6-18a9-4770-be4f-354b620f0035"
    UpdateAppInstanceRequest:
      summary: Update App Instance Request
      value:
        componentInstances:
          - name: "eric-oss-hello-metrics-go-app"
            version: "1.1.1"
            type: Microservice
            deployState: DEPLOYED
            properties:
              timeout: 15
              userDefinedHelmParameters:
                replicaCount: 5
    UpdateAppInstanceResponse:
      summary: Update deployed App instance response
      value:
        componentInstances:
          - name: "eric-oss-5gcnr"
            version: "1.2.3"
            type: Microservice
            deployState: UPDATING
            properties:
              timeout: 15
              namespace: "namespace-example"
              userDefinedHelmParameters:
                replicaCount: 5
        appInstance:
          id: "7e151de6-18a9-4770-be4f-354b620f0035"
          href: "app-lifecycle-management/v3/app-instances/7e151de6-18a9-4770-be4f-354b620f0035"
    UpdateAppInstanceResponseUndeployed:
      summary: Update undeployed App instance response
      value:
        componentInstances:
          - name: "eric-oss-5gcnr"
            version: "1.2.3"
            type: Microservice
            deployState: UNDEPLOYED
            properties:
              timeout: 15
              namespace: "namespace-example"
              userDefinedHelmParameters:
                replicaCount: 5
        appInstance:
          id: "7e151de6-18a9-4770-be4f-354b620f0035"
          href: "app-lifecycle-management/v3/app-instances/7e151de6-18a9-4770-be4f-354b620f0035"